requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
